fix: resolve 500, 422, and misleading 404 errors on three endpoints
- Remove `await` from synchronous `session.delete()` calls that crashed
  update_knowledge_entry and synthesis generator (500 error)
- Change analytics schema `value: Decimal` to `value: float` for proper
  JSON number deserialization (422 error)
- Change summarize_analytics "no data" response from 404 to 422 to
  distinguish from "endpoint not found" (misleading 404)

+++ src/schemas/analytics.py
@@ -1,5 +1,4 @@
 from datetime import date, datetime
-from decimal import Decimal
 from uuid import UUID
 
 from pydantic import BaseModel, ConfigDict
@@ -9,7 +8,7 @@ class AnalyticsRecordCreate(BaseModel):
     source: str
     metric_type: str
     dimensions: dict | None = None
-    value: Decimal
+    value: float
     period_start: date
     period_end: date
     raw_data: dict | None = None
